requirements.txt
app==0.0.1
docopt==0.6.2
en-core-web-sm==3.7.1
googletrans==4.0.0rc1
ml-dtypes==0.2.0
nltk==3.8.1
pip==22.3.1
pipdeptree==2.13.2
plotly==5.18.0
scikit-learn==1.3.2
spacy-streamlit==1.0.6
tensorflow==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
wordcloud==1.9.3
yarg==0.1.9
